+++ examples/nuxt-app/pages/[...slug].vue
@@ -1,9 +1,0 @@
-<template>
-  <TideContentPage />
-</template>
-
-<script>
-export default {
-  name: 'TideWrapper'
-}
-</script>