+++ packages/ripple-tide-media/components/global/TideMediaEmbeddedVideo.vue
@@ -1,0 +1,41 @@
+<template>
+  <TideBaseLayout :site="site" :pageTitle="page.title" :pageDescription="page.description" :pageLanguage="page.lang"
+    :updatedDate="page.changed || page.created">
+    <template #aboveHeader>
+      <slot name="aboveHeader"></slot>
+    </template>
+    <template #primaryNav>
+      <slot name="primaryNav"></slot>
+    </template>
+    <template #breadcrumbs>
+      <slot name="breadcrumbs"></slot>
+    </template>
+    <template #aboveBody="{ hasBreadcrumbs }">
+      <TideMediaHeader :header="page.header" :hasBreadcrumbs="hasBreadcrumbs" />
+    </template>
+    <template #body>
+      <TideMediaBody :media="page.media">
+        <RplMediaEmbed :title="page.title" :src="page.media.url" type="video" class="tide-media__embedded-video" />
+      </TideMediaBody>
+    </template>
+    <template #sidebar>
+      <slot name="sidebar"></slot>
+    </template>
+    <template #footer>
+      <slot name="footer"></slot>
+    </template>
+  </TideBaseLayout>
+</template>
+
+<script setup lang="ts">
+import { TideSiteData } from '@dpc-sdp/ripple-tide-api/types'
+import type { TideMediaPage } from '../../types'
+import { RplMediaEmbed } from '@dpc-sdp/ripple-ui-core'
+
+interface Props {
+  site: TideSiteData
+  page: TideMediaPage
+}
+
+defineProps<Props>()
+</script>