+++ packages/ripple-tide-media/src/index.ts
@@ -1,4 +1,0 @@
-import tideMediaEmbeddedVideoMapping from './pages/embedded-video/embedded-video-mapping.js'
-import tideMediaAudioMapping from './pages/audio/audio-mapping.js'
-
-export default [tideMediaEmbeddedVideoMapping, tideMediaAudioMapping]