+++ packages/ripple-tide-publication/mapping/index.ts
@@ -1,0 +1,4 @@
+import tidePublicationModule from './publication.js'
+import tidePublicationPageModule from './publication-page.js'
+
+export default [tidePublicationModule, tidePublicationPageModule]