+++ packages/ripple-tide-publication/src/index.ts
@@ -1,4 +1,0 @@
-import tidePublicationModule from './mapping/publication.js'
-import tidePublicationPageModule from './mapping/publication-page.js'
-
-export default [tidePublicationModule, tidePublicationPageModule]